ModelLoading finished.
GitIgnore:
ObjLoader, a new folder later used for Model Loading.

Add:
ResourceLoader.h/cpp, included tiny_obj_loader.
ResourceLoader.h/cpp, static MeshData* LoadModel(), returns a filled MeshData with the given model's file directory.

RenderFactory.h/cpp, now included the Pavillion, a model that is included. Makes use of ResourceLoader.

BufferData.h, Vertex, bool operator==() copy operator for the Vertex.
BufferData.h, size_t operator(), the size of the vertex data, using hash functions.

Change:
All indices are now from uint16_t to uint32_t to take in account for the larger indice sizes.

main.cpp, only rendering an object. This will crash the program for everyone because no model is uploaded.

+++ VulkanRenderer/VulkanRendererWindows/MainRenderer/RenderObjects/BaseRenderObject.h
@@ -27,7 +27,7 @@ public:
 	Transform* GetTransform() const { return m_Transform; }
 
 	BufferData<Vertex>* GetVertexData() const { return p_RenderObjectData->m_MeshData->GetVertexData(); }
-	BufferData<uint16_t>* GetIndexData() const { return p_RenderObjectData->m_MeshData->GetIndexData(); }
+	BufferData<uint32_t>* GetIndexData() const { return p_RenderObjectData->m_MeshData->GetIndexData(); }
 
 protected:
 	const size_t m_ID;
